backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# Database (JSON-based for prototype)
tinydb==4.8.0
python-dateutil==2.8.2

# PDF Generation
reportlab==4.0.9
pypdf==4.0.1

# Async & Background Tasks
celery==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Logging & Monitoring
python-json-logger==2.0.7
colorlog==6.8.2

# Utils
python-dotenv==1.0.0
pytz==2024.1
email-validator==2.1.0
phonenumbers==8.13.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.5.1

# Machine Learning & Data Science
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
openpyxl==3.1.2
xlrd==2.0.1
pyarrow==15.0.0
matplotlib==3.8.2
seaborn==0.13.1

# AI/ML (Optional - for future enhancements)
# langchain==0.1.0
# openai==1.10.0
